agents/searcher_agent/pyproject.toml
[project]
name = "searcher-agent"
version = "0.1.0"
description = "Search locations closest to patient and ranks them"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "genai-protocol>=1.0.9",
    "python-dotenv>=1.1.0",
    "groq>=0.30.0",
    "aiohttp",
    "geopy",
    "requests",
]
